Ensure that empty arrays are not treated as scalars
When running:

    @gp Float64[]

the data sent to Gnuplot (stored in $data0) are " Float64[]" rather
than "". This is because the empty array is incorrectly treated as a
scalar. This commit fixes that.

Note that the end result with and without this commit is the same.
Nothing is (re)drawn. Before, gnuplot complained with:

    Bad data on line 1 of file $data0

Now, gnuplot says:

    warning: Skipping data file with no valid points