Switch to using display() to show the plots
This brings better compatibility with the Julia package ecosystem.
Now, if Gnuplot.jl is used in an environment capable of showing
multimedia content (IJulia, VS Code, Pluto), their internal viewer
will take precedence over using gnuplot's built-in viewer. In the
REPL, gnuplot viewer is still used by default.

In VS Code, for example, when the *Use Plot Pane* settings is enabled,
the plots show in that pane, but when it is disabled, gnuplot viewer is
automatically used.

For people who prefer to always use the gnuplot viewer, they can set
Gnuplot.options.gpviewer to true. This should result in the same
behaviour as before this commit.